+++ app/ListingForm/page.tsx
@@ -1,0 +1,242 @@
+"use client";
+import { useState } from "react";
+import { Button } from "@/app/components/ui/button";
+import { Icons } from "@/app/utils/icons";
+
+import { Form } from "react-hook-form";
+import { FormDataSchema } from "@/lib/FormSchema/schema";
+import { z } from "zod";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm, SubmitHandler } from "react-hook-form";
+
+type Inputs = z.infer<typeof FormDataSchema>;
+
+const steps = [
+  {
+    id: "Step 1",
+    name: "item name and image",
+    fields: ["iname", "image"],
+  },
+  {
+    id: "Step 2",
+    name: "Additional Details",
+    fields: ["condition", "Description"],
+  },
+  { id: "step 3", name: "Price", fields: ["Price", "Available till"] },
+  { id: "Step 4", name: "Complete" },
+];
+
+export default function page() {
+  const [isDialogOpen, setIsDialogOpen] = useState(false);
+
+  const [previousStep, setPreviousStep] = useState(0);
+  const [currentStep, setCurrentStep] = useState(0);
+  const delta = currentStep - previousStep;
+
+  const {
+    register,
+    handleSubmit,
+    watch,
+    reset,
+    trigger,
+    formState: { errors },
+  } = useForm<Inputs>({
+    resolver: zodResolver(FormDataSchema),
+  });
+
+  const processForm: SubmitHandler<Inputs> = (data) => {
+    console.log(data);
+    reset();
+  };
+
+  type FieldName = keyof Inputs;
+
+  const next = async () => {
+    const fields = steps[currentStep].fields;
+    const output = await trigger(fields as FieldName[], { shouldFocus: true });
+
+    if (!output) return;
+
+    if (currentStep < steps.length - 1) {
+      if (currentStep === steps.length - 2) {
+        await handleSubmit(processForm)();
+      }
+      setPreviousStep(currentStep);
+      setCurrentStep((step) => step + 1);
+    }
+  };
+
+  const prev = () => {
+    if (currentStep > 0) {
+      setPreviousStep(currentStep);
+      setCurrentStep((step) => step - 1);
+    }
+  };
+
+  return (
+    <>
+      <div>{`Step ${currentStep + 1} of 4.`}</div>
+      <form className="mt-12 py-12" onSubmit={handleSubmit(processForm)}>
+        {currentStep === 0 && (
+          <>
+            <div className="mt-10 grid  gap-x-6 gap-y-8 sm:grid-cols-6">
+              <div className="sm:col-span-3">
+                <label
+                  htmlFor="iname"
+                  className="block text-sm font-medium leading-6 text-gray-900"
+                >
+                  Item Name
+                </label>
+                <div className="mt-2">
+                  <input
+                    type="text"
+                    id="iname"
+                    {...register("iname")}
+                    autoComplete="given-name"
+                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-sky-600 sm:text-sm sm:leading-6"
+                  />
+                  {errors.iname?.message && (
+                    <p className="mt-2 text-sm text-red-400">
+                      {errors.iname.message}
+                    </p>
+                  )}
+                </div>
+              </div>
+              <div className="sm:col-span-3">
+                <label
+                  htmlFor="iname"
+                  className="block text-sm font-medium leading-6 text-gray-900"
+                >
+                  Image
+                </label>
+                <div className="mt-2">
+                  <input
+                    type="file"
+                    id="images"
+                    {...register("images")}
+                    autoComplete="given-name"
+                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-sky-600 sm:text-sm sm:leading-6"
+                  />
+                  {errors.images?.message && (
+                    <p className="mt-2 text-sm text-red-400">
+                      {errors.images.message}
+                    </p>
+                  )}
+                </div>
+              </div>
+            </div>
+          </>
+        )}
+
+        {currentStep === 1 && (
+          <>
+            <div className="mt-10 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
+              <div className="sm:col-span-3">
+                <label
+                  htmlFor="country"
+                  className="block text-sm font-medium leading-6 text-gray-900"
+                >
+                  Condition
+                </label>
+                <div className="mt-2">
+                  <select
+                    id="condition"
+                    {...register("condition")}
+                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:ring-2 focus:ring-inset focus:ring-sky-600 sm:max-w-xs sm:text-sm sm:leading-6"
+                  >
+                    <option>Brand New</option>
+                    <option>Like New</option>
+                    <option>Old</option>
+                  </select>
+                  {errors.condition?.message && (
+                    <p className="mt-2 text-sm text-red-400">
+                      {errors.condition.message}
+                    </p>
+                  )}
+                </div>
+              </div>
+
+              <div className="col-span-full">
+                <label
+                  htmlFor="street"
+                  className="block text-sm font-medium leading-6 text-gray-900"
+                >
+                  Description
+                </label>
+                <div className="mt-2">
+                  <input
+                    type="text"
+                    id="Description"
+                    {...register("Description")}
+                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-sky-600 sm:text-sm sm:leading-6"
+                  />
+                  {errors.Description?.message && (
+                    <p className="mt-2 text-sm text-red-400">
+                      {errors.Description.message}
+                    </p>
+                  )}
+                </div>
+              </div>
+            </div>
+          </>
+        )}
+
+        {currentStep === 2 && (
+          <>
+            <h2 className="text-base font-semibold leading-7 text-gray-900"></h2>
+            <p className="mt-1 text-sm leading-6 text-gray-600">
+              Thank you for your submission.
+            </p>
+          </>
+        )}
+        {currentStep === 3 && <h1>Card</h1>}
+      </form>
+      <div className="mt-8 pt-5">
+        <div className="flex justify-between">
+          <button
+            type="button"
+            onClick={prev}
+            disabled={currentStep === 0}
+            className="rounded bg-white px-2 py-1 text-sm font-semibold text-sky-900 shadow-sm ring-1 ring-inset ring-sky-300 hover:bg-sky-50 disabled:cursor-not-allowed disabled:opacity-50"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              fill="none"
+              viewBox="0 0 24 24"
+              strokeWidth="1.5"
+              stroke="currentColor"
+              className="h-6 w-6"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                d="M15.75 19.5L8.25 12l7.5-7.5"
+              />
+            </svg>
+          </button>
+          <button
+            type="button"
+            onClick={next}
+            disabled={currentStep === steps.length - 1}
+            className="rounded bg-white px-2 py-1 text-sm font-semibold text-sky-900 shadow-sm ring-1 ring-inset ring-sky-300 hover:bg-sky-50 disabled:cursor-not-allowed disabled:opacity-50"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              fill="none"
+              viewBox="0 0 24 24"
+              strokeWidth="1.5"
+              stroke="currentColor"
+              className="h-6 w-6"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                d="M8.25 4.5l7.5 7.5-7.5 7.5"
+              />
+            </svg>
+          </button>
+        </div>
+      </div>
+    </>
+  );
+}
